ADD - seat list & form

+++ Scripts/DB.sql
@@ -48,7 +48,7 @@ CREATE TABLE ROOM(
   seats_number INT(2),
   floor_number INT(2),
   reservation_type INT(2),
-  state INT(2),
+  state TINYINT(1),
   room_type_id INT,
   center_cif VARCHAR(9) NOT NULL,
   CONSTRAINT room_room_type_id_fk FOREIGN KEY (room_type_id) REFERENCES ROOM_TYPE(id) ON UPDATE CASCADE ON DELETE CASCADE,
@@ -68,7 +68,7 @@ CREATE TABLE RESERVATION(
 CREATE TABLE SEAT(
   id INT AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(50),
-  state INT(2),
+  state TINYINT(1),
   room_id INT,
   CONSTRAINT seat_room_id_fk FOREIGN KEY (room_id) REFERENCES ROOM(id) ON UPDATE CASCADE ON DELETE CASCADE
 );
